Add methods to MongoDBPersister and migrate
Migrates the MongoDBPersister to b_pymongo.py for consistent naming
convention.
